Writing an API for the Profile model
