+++ social_network/API/serializers.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+from account.models import *
+
+class Profileserializers(serializers.ModelSerializer):
+    class Meta:Profile
+    fields='__all__'
