+++ social_network/social_network/settings.py
@@ -40,12 +40,15 @@ INSTALLED_APPS = [
     'rest_framework',
     'account',
     'cart',
+    'API',
+    'corsheaders',
 ]
 
 
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
+    'corsheaders.middleware.CorsMiddleware',   # این برای تنظیمات cors API هست
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
@@ -125,3 +128,13 @@ STATIC_URL = 'static/'
 # https://docs.djangoproject.com/en/4.2/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
+
+CORS_ALLOW_ALL_ORIGINS = True  # این برای دسترسی به همهی دامنه هاست
+
+CORS_ALLOW_METHODS = [   # این هم متد هایی هستن که سمت فرانت ازشمون استفاده میشن
+    'GET',
+    'POST',
+    'PUT',
+    'DELETE',
+    'OPTIONS'
+]
