+++ social_network/social_network/urls.py
@@ -33,5 +33,6 @@ urlpatterns = [
     path('reset', views.reset_password , name='reset_password'),
     path('accounts/reset_password/', auth_views.PasswordResetView.as_view(), name='password_reset'),
     path('home/', include('cart.urls')),
+    path('API', include('API.urls'))
 ]+ static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
